[WIP] repo: repodata interface
